Refactor. Move.
